// src/common/enums/shared.enum.ts
export const superAdminEnums = {
  EMAIL: "superAdmin@Amer.com",
  USER_NAME: "superAdmin",
  FULL_NAME: "Super Admin",
  PHONE_NUMBER: "01124284915",
  PASSWORD: "mostafa@222",
  IS_MALE: 1,
  ROLE_EN_NAME: "Super Admin",
  ROLE_AR_NAME: "المدير العام",
  DEPARTMENT_EN_NAME: "Super Admin Department",
  DEPARTMENT_AR_NAME: "قسم المدير العام",
  POSITION_EN_NAME: "Super Admin Position",
  POSITION_AR_NAME: "المدير العام"
}
export const languageEnums = {
  ENGLISH: "en",
  ARABIC: "ar",
};

export const isDeleteStatusEnums = {
  DELETED: 1,
  NOT_DELETED: 0,
}

export const activeStatusEnums = {
  ACTIVE: 1,
  INACTIVE: 0,
}

export const mailerEnums = {
  MAILER_SUBJECT: 'Welcome to MyPorto CRM!',
}

export const autogeneratedPasswordEnums = {
  LENGTH: 8,
}